refactor: move command to match Core command path
